feat: 하단 버튼 컴포넌트 분리
feat: 첫 번째 단계 인증을 useAuth 커스텀 훅으로 분리
