+++ src/components/pages/Coupon.tsx
@@ -19,6 +19,7 @@ import {
 } from "@tanstack/react-table";
 import { ArrowUpDown } from "lucide-react";
 import { useEffect, useState } from "react";
+import NavigationButtons from "../ui/custom/navigationButton";
 
 // 컬럼 정의
 export const columns: ColumnDef<CouponData>[] = [
@@ -220,6 +221,7 @@ export default function Coupon({
         </TableBody>
       </Table>
       <p className="items-center text-2xl">{CouponMessage(loading)}</p>
+      <NavigationButtons prev={onPrev} next={onNext} />
     </div>
   );
 }
