+++ src/hooks/useAuth.tsx
@@ -1,0 +1,32 @@
+import { useEffect, useState } from "react";
+
+
+
+export default function useAuth() {
+
+    const [isAuthenticated, setIsAuthenticated] = useState<boolean | null>(null);
+    
+    useEffect(() => {
+        const checkAuth = async () => {
+          try {
+            const response = await fetch(
+              `${import.meta.env.VITE_API_URL}/auth/check`,
+            //   {
+            //     credentials: "include",
+            //   }
+            );
+            if (!response.ok) {
+              throw new Error("인증 정보 없음.");
+            }
+            setIsAuthenticated(response.status === 200);
+          } catch (error) {
+            console.error("Auth check error:", error);
+            setIsAuthenticated(false);
+          }
+        };
+    
+        checkAuth();
+      }, []);
+
+      return {isAuthenticated};
+}
